Make Service class and logic of CRUD process
Make Service class and 5 methods to handle CRUD request.
Business logics are not shown in controller, they are in service.
It needs to have Repository class for data handling.
